ecto_gss v0.1.0 EctoGSS.Repo

Repository to use Google Spreadsheets as persistence layer for objects.

Types

ecto_object()

ecto_object() :: Ecto.Changeset.t | Ecto.Schema.t

result()

result() :: {:ok, Ecto.Schema.t} | {:error, Ecto.Changeset.t}

state()

state() :: map

Functions

all(schema, opts \\ [])

all(Ecto.Queryable.t, Keyword.t) :: [Ecto.Schema.t]

Get several records:

by range of rows: start_id and end_id options;

by exact list of rows: rows option.

delete(record)

delete(ecto_object) :: result

Delete an existing record.

delete!(record)

delete!(ecto_object) :: result | no_return

Delete an existing record, raise in case of error.

get(schema, id)

get(Ecto.Queryable.t, integer) ::
  Ecto.Schema.t |
  nil |
  no_return

Get a record by row id, raise if not found.

get!(schema, id)

get!(Ecto.Queryable.t, integer) :: Ecto.Schema.t | no_return

Get a record by row, raise if not found.

insert(changeset)

insert(ecto_object) :: result

Add a new record.

insert!(changeset)

insert!(ecto_object) :: Ecto.Schema.t | no_return

Add a new record, raise in case of error.

insert_or_update(changeset)

insert_or_update(Ecto.Changeset.t) :: result

Update an existing record, or insert a new one.

insert_or_update!(changeset)

insert_or_update!(Ecto.Changeset.t) :: Ecto.Schema.t | no_return

Update an existing record, or insert a new one, raise in case of error.

start_link()

start_link() :: {:ok, pid}

update(changeset)

update(ecto_object) ::
  {:ok, Ecto.Schema.t} |
  {:error, Ecto.Changeset.t}

Update an existing record with a row id.

update!(changeset)

update!(Ecto.Changeset.t) :: Ecto.Schema.t | no_return

Update an existing record, raise in case of error.